Build tests and benchmarks locally and add separate shared build-flags
Naturally, given tests and benchmark link against the built libgrapheme,
it makes no sense to compile them in the local build architecture.

Additionally, we add the establisheed SHFLAGS- and SOFLAGS-variables to
config.mk to give finer control over build-flags specific for the
shared object files and linking of the final so. This saves us from
indiscriminately compiling all object files as PIC and setting soname
for the test- and benchmark-binaries, which makes zero sense.

Signed-off-by: Laslo Hunhold <dev@frign.de>