Properly return offset when input is shorter or equal to one codepoint
I sadly didn't catch this bug with automatic testing, given I didn't
cover the shorthand-check at the beginning of the breakpoint-detection
functions in any test-case. Additionally, it would be shadowed when
simply working with UCS-4-arrays.

On a higher level, this surfaced when checking the title-case of
one-character-strings, given it would first get the next word-break
(which would be underreported as 1 instead of something >1) only to
subtract the real offset of the grapheme-cluster-bytelen later on,
leading to an underflow and infinite loop.

Thanks to polarisFuton9719 for reporting this bug!

Signed-off-by: Laslo Hunhold <dev@frign.de>