Add manual-template-engine and manuals for the case-conversion functions
Given the undeniable redundancy within the manuals it makes a lot of
sense to have a simple template-engine for generating the manual files
on the fly.

After fiddling with m4 and finding out for the third time that it just
hasn't aged well I realized that it makes most sense to simply use
heredocs within small shell-scripts.

Signed-off-by: Laslo Hunhold <dev@frign.de>