Move $VERSION from config.mk to Makefile and group variables
Same as with $MAN_DATE and $UNICODE_VERSIONS, which are parameters that
are not really meant to be changed from the outside depending on the
build system/architecture, it does not make much sense to place the
VERSION variable explicitly within config.mk even though this has been
the usual approach in suckless projects (cargo-cult?).

Given the order of specification it will still always be possible to
override VERSION from config.mk.

Signed-off-by: Laslo Hunhold <dev@frign.de>