Add unit tests for all segmentation functions
Now all functions in the library are covered by exhaustive unit tests
which supplement the already present conformance tests to make sure
that the thin layer between API and implementation is also working as
expected.

At this point I would assess that libgrapheme is a stable foundation
for using it in the real world and now preparation can go underway
to prepare the release of version 2.

Signed-off-by: Laslo Hunhold <dev@frign.de>