Add ./configure-script with presets for common systems
After quite a few requests and a bit of reflection on my behalf I've
decided to add a very simple ./configure-script that automatically
modifies config.mk to make it fit for common systems.

Even though it's reasonable to simply have out-commentable options
in the config.mk, it is admittedly more convenient to have such a
script available, especially to accomodate more systems along the way.

uname(1) is Posix compliant and this ./configure-script is in no way
comparable to the horrible autoconf-insanity and won't take an eternity
to run. It's also completely optional and merely a
quality-of-life-addition for those working with libgrapheme manually.

Signed-off-by: Laslo Hunhold <dev@frign.de>