Do not falsely read entire buffer instead of simply the filled with
This was caught via dynamic analysis (clang asan), which I can definitely
recommend.

Rust evangelists might see this as a prime example for why C is bad, but
I still think the benefits outweigh the risks if you consider the
maturity of tooling to catch these kinds of errors. In an ideal world we
would all be programming in Ada, but C's portability is unmatched.

Signed-off-by: Laslo Hunhold <dev@frign.de>