Replace all POSIX-features to become fully ISO-C99
As it turned out, the only things that needed replacing were
getline(), strdup() and the timing in the benchmarks. Analogously,
we replace -D_DEFAULT_SOURCE with -D_ISOC99_SOURCE.

This way we further extend the number of platforms where libgrapheme can
be compiled and run on, e.g. MSVC still does not include getline().

Signed-off-by: Laslo Hunhold <dev@frign.de>