Update grapheme break algorithm to Unicode version 15.1.0
While the change to the algorithm looks harmless in the specification,
it comes at the price of more complexity because we have to keep track
of a relatively complex state for a sequence of indic conjunct breaks.

Fortunately adding so many additional classes only decreases the
compression ratio for the grapheme cluster LUTs by ~0.5%.

We now pass all 1187 character tests.

Signed-off-by: Laslo Hunhold <dev@frign.de>